Towards #128 Test folder structure components/charts.
