+++ test/component/barsVerticalTest.js
@@ -1,5 +1,5 @@
-let d3Ez = require("../");
-let tape = require("tape");
+let test = require("tape");
+let d3Ez = require("../../build/d3-ez");
 let jsdom = require("jsdom");
 let d3 = require("d3");
 let fs = require("fs");
@@ -23,35 +23,31 @@ function readSvgFile(file, element) {
 	return element.insertAdjacentHTML("beforeend", str);
 }
 
-tape("setup", function(t) {
+test("setup", function(t) {
 	let JSDOM = jsdom.JSDOM;
 	global.document = new JSDOM().window.document;
 	t.end();
 });
 
-tape("componentBarsStackedTest", function(t) {
+test("componentBarsVerticalTest", function(t) {
 	let width = 300;
 	let height = 300;
-	let radius = 150;
-	let innerRadius = 10;
 
 	// Load 'exected' svg file into first div.
 	let actualDiv = document.createElement("div");
-	readSvgFile("./test/svg/componentBarsCircular.svg", actualDiv);
+	readSvgFile("./test/component/svg/barsVertical.svg", actualDiv);
 
 	// Construct 'actual' svg using d3-ez component.
 	let expectedDiv = document.createElement("div");
 
-	let myChart = d3Ez.component.barsCircular()
-		.radius(radius)
-		.innerRadius(innerRadius);
+	let myChart = d3Ez.component.barsVertical()
+		.width(width)
+		.height(height);
 
 	let chartHolder = d3.select(expectedDiv);
 	chartHolder.append("svg")
 		.attr("width", width)
 		.attr("height", height)
-		.append('g')
-		.attr('transform', 'translate(' + radius + ',' + radius + ')')
 		.datum(data)
 		.call(myChart);
 
@@ -60,8 +56,7 @@ tape("componentBarsStackedTest", function(t) {
 		let actual = actualDiv.getElementsByTagName("svg")[0].innerHTML;
 		let expected = expectedDiv.getElementsByTagName("svg")[0].innerHTML;
 
-		// Issue with slightl different arc path values on TravisCI vs MacBook
-		// t.equal(actual, expected);
+		t.equal(actual, expected);
 		t.end();
 	}, 600);
 });
