+++ test/component/html/barsCircular.html
