+++ test/component/html/barsVertical.html
