+++ test/component/html/legend.html
@@ -4,8 +4,8 @@
 <head>
   <title>d3-ez : Legend Test</title>
   <script src="https://d3js.org/d3.v5.min.js"></script>
-  <script src="../../build/d3-ez.js"></script>
-  <link rel="stylesheet" type="text/css" href="../../build/d3-ez.css" />
+  <script src="../../../build/d3-ez.js"></script>
+  <link rel="stylesheet" type="text/css" href="../../../build/d3-ez.css" />
 </head>
 
 <body>
