+++ test/component/roseChartSectorTest.js
@@ -1,5 +1,5 @@
-let d3Ez = require("../");
-let tape = require("tape");
+let test = require("tape");
+let d3Ez = require("../../build/d3-ez");
 let jsdom = require("jsdom");
 let d3 = require("d3");
 let fs = require("fs");
@@ -23,13 +23,13 @@ function readSvgFile(file, element) {
 	return element.insertAdjacentHTML("beforeend", str);
 }
 
-tape("setup", function(t) {
+test("setup", function(t) {
 	let JSDOM = jsdom.JSDOM;
 	global.document = new JSDOM().window.document;
 	t.end();
 });
 
-tape("componentRoseChartSectorTest", function(t) {
+test("componentRoseChartSectorTest", function(t) {
 	let width = 300;
 	let height = 300;
 	let radius = 300;
@@ -38,7 +38,7 @@ tape("componentRoseChartSectorTest", function(t) {
 
 	// Load 'exected' svg file into first div.
 	let actualDiv = document.createElement("div");
-	readSvgFile("./test/svg/componentRoseChartSector.svg", actualDiv);
+	readSvgFile("./test/component/svg/roseChartSector.svg", actualDiv);
 
 	// Construct 'actual' svg using d3-ez component.
 	let expectedDiv = document.createElement("div");
