+++ test/indexTest.js
@@ -1,21 +1,21 @@
+let test = require("tape");
 let d3Ez = require("../");
-let tape = require("tape");
 let jsdom = require("jsdom");
 
-tape("setup", function(t) {
+test("setup", function(t) {
 	let JSDOM = jsdom.JSDOM;
 	global.document = new JSDOM().window.document;
 	t.end();
 });
 
-tape("indexTest", function(t) {
+test("indexTest", function(t) {
 	t.equal(d3Ez.author, "James Saunders", "Returns author");
 	t.equal(d3Ez.license, "GPL-2.0", "Returns license");
 
 	t.end();
 });
 
-tape("titleTest", function(t) {
+test("titleTest", function(t) {
 	let title = d3Ez.component.title().mainText("Foo").subText("Bar");
 	t.equal(title.mainText(), "Foo", "Returns main title");
 	t.equal(title.subText(), "Bar", "Returns sub-title");
@@ -23,7 +23,7 @@ tape("titleTest", function(t) {
 	t.end();
 });
 
-tape("legendTest", function(t) {
+test("legendTest", function(t) {
 	let legend = d3Ez.component.legend();
 
 	legend.title("Foo Bar");
@@ -32,7 +32,7 @@ tape("legendTest", function(t) {
 	t.end();
 });
 
-tape("paletteTest", function(t) {
+test("paletteTest", function(t) {
 	let expected1 = ["#5da5da", "#faa43a", "#60bd68", "#f17cb0", "#b2912f", "#b276b2", "#decf3f", "#f15854", "#4d4d4d"];
 	let actual1 = d3Ez.palette.categorical(1);
 	t.deepEqual(actual1, expected1, "Categorical palette");
